+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Download current and previous releases from [https://github.com/ludus-leonis/Fre
 ### Other changes
 
 * added automated tests for zip/tgz packages after build
+* added setup variant variant grid-hex2 for hexes rotated 90°
 * fixed docker build not pulling latest php baseimage
 * refactored api tests in all/latest